Add files via upload
In this file, I have wrote a source code for designing Home page and Transaction page for the Sparks bank. I have created a buttons on a home page as a options. Used splash on the home page as Bank Icon. In the Transaction page I have created a table which contains information related to the customer.